Add files via upload
HashCode Generator:
This Java program implements a simple hash function to generate a hash value based on the ASCII sum of characters in a string. It multiplies the sum by a user-defined hash set size and returns the final hash value. It demonstrates basic hashing principles commonly used in hash tables and hash sets.